Function References

Core

SingleLinkagePartitions.construct_partition — Function

construct_partition(s::SLINKState, num_fuses::Integer)

This function allocates. Run slink! on s first. Return type is Memory{Memory{Int}}.

num_fuses identifies the partition tree level. It can take on values between 0 to get_cardinality(s).

SingleLinkagePartitions.construct_partition_tree — Function

construct_partition_tree(s::SLINKState)

This function allocates. Run slink! on s first. Return type is Memory{Memory{Memory{Int}}}.

SingleLinkagePartitions.iterated_sl — Function

iterated_sl(
    level_config::UseMaxDeviation,
    dist_callable,
    X0::Union{AbstractVector{<:AbstractVector{T}}, AbstractVector{<:AbstractVector{Complex{T}}}};
    acceptance_factor::T = convert(T, 0.7),
    max_iter::Int = 100,
) where {T <: AbstractFloat}

Returns a partition P, and iters_ran. Return type is Tuple{Vector{Vector{Int}}, Int}. Iteratively run single-linkage clustering with UseMaxDeviation and picklevel( to select a partition at every iteration. Some of the parts in the selected partition are chosen to be moved to a solution partition, and the associated points won't be used for the next iteration of single-linkage clustering. See the documentation website for details on the iteration algorithm, including terminating conditions.

SingleLinkagePartitions.slink! — Function

slink!(s::SLINKState{T}, eval_dist, X::AbstractMatrix) where {T <: AbstractFloat}

X is the point cloud matrix, and must be 1-indexing with stride 1. X[d,n] should be the d-th coordinate dimension, n-th data entry. Returns nothing.

eval_dist should be a callable (struct or function) that takes inputs (x::AbstractVector{T}, y::AbstractVector{T}) and outputs a number of type T. An example is EuclideanDistance.

slink!(s::SLINKState{T}, eval_dist, X::AbstractVector{<:AbstractVector}) where {T <: AbstractFloat}

SingleLinkagePartitions.SLINKState — Type

SLINKState(::Type{T}, N::Integer) where {T <: AbstractFloat}

Instantiate a new SLINKState{T} buffer for N observations. The internals of SLINKState are considered private API.

Utilities

SingleLinkagePartitions.construct_distances — Function

construct_distances(s::SLINKState)

Returns the set of minimum distanes between consecutive nested partitions in the partition tree associated with s. Run slink! on s first.

This function allocates. The returned array should have get_cardinality(s)-1 entries. It is of type Memory{T} if s::SLINKState{T}.

SingleLinkagePartitions.compute_max_deviation — Function

compute_max_deviation(X, s::SLINKState, level::Integer)

This version extracts the partition assocaited with level from pt, then calls compute_max_deviation(X, partition).

compute_max_deviation(
    X::Union{AbstractVector{<:AbstractVector{T}}, AbstractVector{<:AbstractVector{Complex{T}}}},
    partition::AbstractVector{<:AbstractVector{Int}},
) where T <: AbstractFloat

Given a point set X on a D-dimensional vector space, we define the maximum deviation of a partition P_X of X as:

ρ = maximum(
    maximum(
        maximum(
            abs(a[d]-mean(A)[d])
            for a in A
        )
        for A in P_X
    )
    for d in 1:D
)

Computes the maximum of the maximum deviations of a partition of X. This is ρ_dim(X[partition]) from the documentation website home page.

Returns a number of type T.

SingleLinkagePartitions.EuclideanDistance — Type

struct EuclideanDistance end

Callable struct:

(::EuclideanDistance)(x::AbstractVector, y::AbstractVector)

Returns the Euclidean distance between x and y, without allocating.

SingleLinkagePartitions.get_cardinality — Function

get_cardinality(s::SLINKState)

Returns the number of entries for the input set. Return type is Int.

SingleLinkagePartitions.pick_level — Function

pick_level(C::UseSLDistance, s::SLINKState, unused_args...)

See UseSLDistance for a description of merging distance. Returns the partition for first level l-1 such that the merging distance between levels $l$ and $l-1$ is larger than the C.atol struct field. The search starts from $l$ = 1, $l-1$ = 0 (the leaf level).

This function allocates. Run slink! on s first. Return type is Int.

pick_level(C::UseCumulativeSLDistance, s::SLINKState, unused_args...)

See UseCumulativeSLDistance for a description of merging distance. Returns the partition for level l-1 such that the cumulative sum of merging distances from the level 0 (the leaf level) to level lis larger than the C.atol struct field. The search starts from l = 1, l-1 = 0.

This function allocates. Run slink! on s first. Return type is Int.

pick_level(
    C::UseMaxDeviation{T},
    s::SLINKState{T},
    X::AbstractVector{<:AbstractVector{<:Number}},
) where {T}

Returns the level index, of type Int, from the partition tree such that its maximum deviation is approximately a best match to the maximum deviation threshold recorded in C.

This method uses a bracketing binary search between levels 0 and get_cardinality(s)-1. See UseMaxDeviation for further details on the terminology.

SingleLinkagePartitions.get_max_level — Function

get_max_level(s::SLINKState)

Returns the maximum level of the partition tree that s represents. Run slink! on s first.

SingleLinkagePartitions.vecs_to_mat — Function

vecs_to_mat(X::AbstractVector{<:AbstractVector{T}}) where {T}

Returns a matrix constructed from X, with X[d,n] == X[n][d]. Return type is Matrix{T}.

Application routines

SingleLinkagePartitions.avg_duplicates — Function

avg_duplicates(X::AbstractVector, y::AbstractVector, zero_tol::AbstractFloat)

For any locations in X that are nearly identical (with respect to zero_tol), reduce the corresponding y entries by mean. Use replace_duplicates if the points to be merged should be determined by a score-based approach.

zero_tol is the single-linkage distance threshold. It is similar to a tolerance for merging the points.

Returns

Xc: a reduced version of X.

yc: a reduced version of y.

SingleLinkagePartitions.reduce_pts — Function

reduce_pts(
    level_trait::LevelOption,
    center_trait::PartRepOption,
    dist_callable,
    X::AbstractVector{<:AbstractVector{NT}},
) where {NT <: Number}

Returns:

centers::Memory{XT}, if X::AbstractVector{XT}. This is the set of reduced points.

variances::Memory{NT}: the variances of the reduced points. If there were no points reduced, then the variance entry is zero.

partition::Memory{Memory{Int}}: the partition of X that was used to make the reduction. The elements of a part are replaced with a part representative according to center_trait.

reduce_pts(
    level_trait::LevelOption,
    center_trait::PartRepOption,
    dist_callable,
    X::Union{AbstractVector{<:AbstractVector{T}}, AbstractVector{<:AbstractVector{Complex{T}}}},
    y::AbstractVector{YT},
) where {T <: AbstractFloat, YT <: Number}

X is the parimary input point set, which is used to determine the reduction. y is a companion input point set, that is reduced according to how points in X are reduced.

Returns:

Xc::Memory{XT}, if X::AbstractVector{XT}. This is the set of reduced points from X.

Xv::Memory{NT}: the variances of the merged points from X. If there were no points merged, then the variance entry is zero.

-yc::Memory{XT}, if X::AbstractVector{XT}. This is the set of reduced points from y.

yv::Memory{NT}: the variances of the merged points from y. If there were no points merged, then the variance entry is zero.

partition::Memory{Memory{Int}}: the partition of X and y that was used to make the reduction. The elements of a part are replaced with a part representative according to center_trait.

reduce_pts(
    level_trait::LevelOption,
    center_trait::PartRepOption,
    dist_callable,
    X::AbstractVector{<:AbstractVector{<:Number}},
    Y::AbstractVector{<:AbstractVector{<:Number}},
)

X is the parimary input point set, which is used to determine the reduction. y is a companion input point set, that is reduced according to how points in X are reduced.

Returns:

Xc::Memory{XT}, if X::AbstractVector{XT}. This is the set of reduced points from X.

Xv::Memory{NT}: the variances of the merged points from X. If there were no points merged, then the variance entry is zero. NT is the underlying floating-point type of X.

-Yc::Memory{XT}, if X::AbstractVector{XT}. This is the set of reduced points from Y.

Yv::Memory{NT}: the variances of the merged points from Y. If there were no points merged, then the variance entry is zero. NT is the underlying floating-point type of Y.

partition::Memory{Memory{Int}}: the partition of X and y that was used to make the reduction. The elements of a part are replaced with a part representative according to center_trait.

SingleLinkagePartitions.replace_duplicates — Function

replace_duplicates(
    copy_trait::CopyTrait,
    score_trait::ScoreTrait,
    X::AbstractVector,
    y::AbstractVector{<:AbstractFloat},
    zero_tol::AbstractFloat,
)

For any locations in X that are nearly identical (with respect to zero_tol), reduce the corresponding y entries by the behavior specified by score_trait. Use avg_duplicates if the mean of the points to be merged should be used instead of a score-based approach.

zero_tol is the single-linkage distance threshold. It is similar to a tolerance for merging the points.

Returns

Xc: a reduced version of X.

yc: a reduced version of y.

Options

Level

SingleLinkagePartitions.LevelOption — Type

abstract type LevelOption end

A trait that is a subtype of LevelOption is used to pick a level of the partition tree, given a tolerance. Use subtypes to see the implemented traits.

SingleLinkagePartitions.UseCumulativeSLDistance — Type

UseCumulativeSLDistance(atol::T) where T <: AbstractFloat

Creates a variable of type UseCumulativeSLDistance{T} <: LevelOption.

Given a partition tree, the part/cluster that merged as we traverse from level l to level l+1 is the merging distance of the two levels.

UseCumulativeSLDistance specifies the threshold on the cumulative merging distance from level 0 to l. For use with reduce_pts. The internals of UseCumulativeSLDistance are considered private API.

SingleLinkagePartitions.UseSLDistance — Type

UseSLDistance(atol::T) where T <: AbstractFloat

Creates a variable of type UseSLDistance{T} <: LevelOption.

Given a partition tree, the part/cluster that merged as we traverse from level l to level l+1 is the merging distance of the two levels.

UseSLDistance specifies the threshold merging distance of level l and l+1. For use with reduce_pts. The internals of UseSLDistance are considered private API.

SingleLinkagePartitions.UseMaxDeviation — Type

UseMaxDeviation(max_dev::T, max_iters::Int, search_convergence_atol::T) where {T <: AbstractFloat}

Creates a variable of type UseMaxDeviation{T} <: LevelOption.

Terminology: Given a point set X on a D-dimensional vector space, we define the maximum deviation of a partition P_X of X as:

ρ = maximum(
    maximum(
        maximum(
            abs(a[d]-mean(A)[d])
            for a in A
        )
        for A in P_X
    )
    for d in 1:D
)

If we order the partitions of a partition tree such that the next partition merges two of the parts in the current partition, i.e. in a nesting order, then the sequence of maximum deviations of the ordered partitions is approximately monotonically increasing. This is because a partition with larger sized parts (partitions in the latter parts of the sequence) in general have more variability within their larger sized parts than a partition with smaller sized parts (partitions in the beginning parts of the sequence).

Description: The UseMaxDeviation trait/config data type specifies that we should choose a partition from a partition tree such that the chosen partition has a maximum deviation that is approximately the best match to a user-specified target value, which is stored as the max_dev struct field.

The reason the implemented methods associated with UseMaxDeviation does not guaranteed to return the best match is because the sequence of maximum deviations that corresponds to the nesting order of partitions is not guaranteed to be monotonic.

The max_dev input should be a non-negative number. It is clamped to zero if the max_dev provided is a negative number. Setting this to 0 corresponds to selecting the trivial partition where every point is a singleton part, i.e., level 0 of the partition tree.

The internals of UseMaxDeviation is considered private API.

UseMaxDeviation(max_dev::T) where {T <: AbstractFloat}

A convenience constructor for UseMaxDeviation(max_dev, 100, max_dev/100).

Partition representative

SingleLinkagePartitions.PartRepOption — Type

abstract type PartRepOption end

A trait that is a subtype of PartRepOption dictates how reduce_pts select the reduced point output from all the points of a given part that is to be reduced. Use subtypes to see the implemented traits.

SingleLinkagePartitions.UseProximitytoMean — Type

struct UseProximitytoMean <: PartRepOption end

This is a trait that specifies reduce_pts should use the element closest to the mean of all the elements in the part to the reduced.

SingleLinkagePartitions.UseMean — Type

struct UseMean <: PartRepOption end

This is a trait that specifies reduce_pts should use the mean of the part elements as the reduced point. This means the output is in general not an element of the part to be reduced.

Score options

SingleLinkagePartitions.UseMinimum — Type

struct UseMinimum <: ScoreTrait end

This is a trait that specifies reduce_pts should use the element that corresponds to the minimum score in the score array.

SingleLinkagePartitions.UseMaximum — Type

struct UseMaximum <: ScoreTrait end

This is a trait that specifies reduce_pts should use the element that corresponds to the maximum score in the score array.

SingleLinkagePartitions.UseScore — Type

UseScore(
    copy_trait::CT,
    score_trait::ST,
    score::AbstractVector{T},
) where {T <: AbstractFloat, ST <: ScoreTrait, CT <: CopyTrait}

Creates a variable of type UseScore{T, ST, CT} <: PartRepOption. The internals of UseScore are considered private API.

SingleLinkagePartitions.ScoreTrait — Type

abstract type ScoreTrait end

A trait that is a subtype of ScoreTrait dictates how reduce_pts uses an extern score array, with eleemnts corresponding to the input point cloud that generated the partition tree, to select the reduced point output from all the points of a given part that is to be reduced. Use subtypes to see the implemented traits.