Function References

Core

SingleLinkagePartitions.geteuclideanmetric — Function

geteuclideanmetric()

Convinence constructor for a DistancesjlMetric type that SingleLinkagePartitions utilizes. returns DistancesjlMetric(Distances.Euclidean())

Partition construction

Dispatch types/traits/configs

SingleLinkagePartitions.UseMaxDeviation — Type

struct UseMaxDeviation{T <: AbstractFloat} <: LevelOption
    max_dev::T # a positive number that specifies a level from a partition tree. For use with `picklevel`.

    # the following are tuning parameters for use with `iteratedsl`.
    max_iters::Int
    search_convergence_atol::T
    verbose::Bool
end

Dispatch type/trait, for use with picking a level via picklevel or iterated single-linkage via iteratedsl.

Terminology: Given a point set X on a D-dimensional vector space, we define the maximum deviation of a partition P_X of X as:

ρ = maximum(
    maximum(
        maximum(
            abs(a[d]-mean(A)[d])
            for a in A
        )
        for A in P_X
    )
    for d in 1:D
)

If we order the partitions of a partition tree such that the next partition merges two of the parts in the current partition, i.e. in a nesting order, then the sequence of maximum deviations of the ordered partitions is approximately monotonically increasing. This is because a partition with larger sized parts (partitions in the latter parts of the sequence) in general have more variability within their larger sized parts than a partition with smaller sized parts (partitions in the beginning parts of the sequence).

Description: The UseMaxDeviation trait/config data type specifies that we should choose a partition from a partition tree such that the chosen partition has a maximum deviation that is approximately the best match to a user-specified target value, which is stored as the max_dev struct field.

The reason the implemented methods associated with UseMaxDeviation does not guaranteed to return the best match is because the sequence of maximum deviations that corresponds to the nesting order of partitions is not guaranteed to be monotonic.

SingleLinkagePartitions.UseSLDistance — Type

struct UseSLDistance{T} <: LevelOption
    atol::T
end

Terminology: The partitions of two consecutive levels in any partition tree pt is such that the larger level corresponds to a larger partition that contains the partition that corresponds to the smaller level. The part/cluster that merged as we traverse from level $l$ to level $l+1$ is the merging distance of the two levels.

Description: UseSLDistance specifies the threshold merging distance of level l and l+1. For use with picklevel(.

SingleLinkagePartitions.UseCumulativeSLDistance — Type

struct UseCumulativeSLDistance{T} <: LevelOption
    atol::T
end

See UseSLDistance for a description on merging distances.

UseCumulativeSLDistance specifies the threshold on the cumulative merging distance from level 0 to l. For use with picklevel(.

Functions

SingleLinkagePartitions.computesl — Function

computesl(
    metric::Union{DistancesjlMetric, InnerProductNorm},
    X::Union{Vector{Vector{T}}, Vector{Vector{Complex{T}}}},
    )::PartitionTree{T} where T <: AbstractFloat

Computes and returns the single-linkage partition tree of X, with respect to distance metric, metric. The partitions in the tree are specified by a level index from 0 to maxlevel(pt), where pt is the returned PartitionTree variable.

See getpartition to select a partition by its level from the return partition tree.

SingleLinkagePartitions.picklevel — Function

picklevel(C::UseSLDistance, pt::PartitionTree{T}, args...)::Int where T <: AbstractFloat

See UseSLDistance for a description of merging distance. Returns the partition for first level l-1 such that the merging distance between levels $l$ and $l-1$ is larger than the C.atol struct field. The search starts from $l$ = 1, $l-1$ = 0 (the leaf level).

picklevel(C::UseCumulativeSLDistance, pt::PartitionTree{T}, args...)::Int where T <: AbstractFloat

See UseSLDistance for a description of merging distance. Returns the partition for level l-1 such that the cumulative sum of merging distances from the level $0$ (the leaf level) to level $l$is larger than the C.atol struct field. The search starts from $l$ = 1, $l-1$ = 0.

picklevel(
    C::UseMaxDeviation{T},
    pt::PartitionTree{T},
    X,
)::Int where T <: AbstractFloat

Returns an approximate best match partition from pt such that its maximum deviation is approximately a best match to C.max_dev.

Use a bracketing binary search between levels 0 and getmaxlevel(pt). See UseMaxDeviation for further details on the terminology.

SingleLinkagePartitions.computemaxdeviation — Function

computemaxdeviation(
    X::Union{Vector{Vector{T}}, Vector{Vector{Complex{T}}}},
    pt::PartitionTree,
    level::Integer,
) where T <: AbstractFloat

This version extracts the partition assocaited with level from pt, then calls computemaxdeviation(X, partition).

computemaxdeviation(
    X::Union{Vector{Vector{T}}, Vector{Vector{Complex{T}}}},
    partition::Vector{Vector{Int}},
) where T <: AbstractFloat

Given a point set X on a D-dimensional vector space, we define the maximum deviation of a partition P_X of X as:

ρ = maximum(
    maximum(
        maximum(
            abs(a[d]-mean(A)[d])
            for a in A
        )
        for A in P_X
    )
    for d in 1:D
)

computemaxdeviation computes ρ a bit more efficiently, with '''julia P_X = collect( X[partition[k]] for k in eachindex(partition) ) '''

SingleLinkagePartitions.iteratedsl — Function

iteratedsl(
    level_config::UseMaxDeviation,
    metric::MetricType,
    X0::Union{Vector{Vector{T}}, Vector{Vector{Complex{T}}}};
    acceptance_factor::T = convert(T, 0.7),
    max_iter = 100,
)::Tuple{Vector{Vector{Int}}, Int} where T <: AbstractFloat

Returns a partition P, anditers_ran. Iteratively run single-linkage clustering withUseMaxDeviation andpicklevel(` to select a partition at every iteration. Some of the parts in the selected partition are chosen to be moved to a solution partition, and the associated points won't be used for the next iteration of single-linkage clustering. See the documentation website for details on the iteration algorithm, including terminating conditions.

Get/Set routines

Partition tree data type PartitionTree

SingleLinkagePartitions.getmaxlevel — Function

getmaxlevel(pt::PartitionTree)::Int

Returns the maximum level of the partition tree.

A valid level for pt can take integer values in [0, getmaxlevel(pt)]. A level indexes a partition in pt.

SingleLinkagePartitions.getpartition — Function

getpartition(pt::PartitionTree, level::Integer)::Vector{Vector{Int}}

level indicates the number of merges that occured to produce a partition. Think of it as an index from 0 to max_level = getmaxlevel(pt).

We use the terminology part to mean an element (i.e. a subset) of a partition. Some folks call it a cluster.

level == 0 corresponds to the leaf level where each point is assigned its own part (i.e. cluster), i.e., we have only singleton parts.

The maximum level, level == getmaxlevel(pt), corresponds to the root level where all points are assigned to be in the same part.

SingleLinkagePartitions.getdistances — Function

getdistances(pt::PartitionTree{T})::Vector{T} where T <: AbstractFloat

Returns the set of minimum distanes between consecutive nested partitions. If pt was constructed from N points, then there are N-1 entries in the returned 1-D array of distances.

SingleLinkagePartitions.generateallpartitions — Function

generateallpartitions(pt::PartitionTree)::Vector{Vector{Vector{Int}}}

returns the set of all partitions in the partition tree, partition_set If N points was used to generate the partition tree pt, then:

partition_set has N+1 entries. Each entry is a partition.

Given the input point set that generated pt, call it X::Vector{Vector{T}}, a partition index 1 <= m <= N+1, part index k, part member index i, if we assign z = partition_set[m][k][i], then X[z] is point that belongs to the m-th partition, k-th part/cluster in the partition, and it is the i-th point in that part/cluster.

SingleLinkagePartitions.getlargestdistance — Function

getlargestdistance(pt::PartitionTree{T})::Vector{T} where T <: AbstractFloat

Returns the largest merge/single-linkage distance.